requirements.txt
asgiref==3.9.1
beautifulsoup4==4.12.3
boto3==1.26.137
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-allauth==65.10.0
django-environ==0.12.0
django-storages==1.13.2
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
Pillow==10.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
